Replace reference of init pid_ns with init pid_t in dev_ns
This closes Cellrox/devns-patches#4

The dev_ns is coupled to pid_ns at creation time. There is no need for
dev_ns to keep a reference to pid_ns. The reference was used to be able
to tell the pid of the respective init task for use by printk; instead,
store the pid_t itself in dev_ns.

Signed-off-by: Oren Laadan <orenl@cellrox.com>
Reported-by: Jeremy Andrus <jeremya@cs.columbia.edu>